Finalize character endpoint
Two new parameter have been added to the character lookup: all_data
and poll. The all_data parameter will toggle between the standard lookup
and a lookup with all of the data flags. The poll parameter will
continuously poll the API until the character has been cached. A new
type of request has been created to handle this continuous polling. It
also handles the various errors states that may be returned and throws a
custom exception for each state. Adds the character update and delete
endpoints.